QueryParser: Set milliseconds of parsed Dates to 0 for consistency
Previously, since we use new Date() here, the milliseconds were set to whatever
they happened to be when new Date() ran. Most of the time that doesn't matter,
but if we did a query like eg. timestamp=2011-8-2 we'd end up with a Date range
of a few milliseconds into 2011-08-02 to a few milliseconds into 2011-08-03.

Then if we construct the same day with new Date(2011, 7, 2) (months are 0-based
in the constructor) we get milliseconds of 0, which can be outside the parser's
range!