New implementation of copyFrom.
Takes fast path code form initArgs when copying from a simple map.

Avoids copying default values and expressions if the two objects are the
same class.

If the objects are different classes of FObject, then only copy
properties that match.

If the source is some unknown type of object, copy any values that match
local properties which are not undefined.