Olympics: Add Polymer MD version of the Olympic demo.
Moves the demo to src/com/google/foam/demos/olympics to be with the others.

You need to run `bower update` in the olympics directory before it will work,
and you need to use its custom HTML page: polymer.html.

The Polymer-powered version is separate from the original, to help demonstrate
the changes necessary to add Polymer to an existing app.