Whole bunch of comics refactoring.
Improve separation between model and controller.

DAOController and others is the MVC Model of DAOControllerView.

DAOControllerView is both a view and a controller for viewing a
DAOController.

A DAO is an MVC model for TableView and others, we adapt a DAO to a
DAOController when we want to browse a DAO using the Browser.