Sink interface change: put, remove args: (sub, obj) => (obj, sub)
Towards #301. 25abf392efdc6db3d5b4e303ac771bed19d68748 changed the
`Sink.put()` and `Sink.remove()` to accept a subscription object as the first
paramter. The idea was to make these methods closer to the `Listener`
interface. Subsequent discussions built a consensus that similarity to
`Listener` is less important that putting the most-used paramter first (and
not breaking existing consumers of `Sink`). This change effectively reverts
the "change the Sink interface" parts of
25abf392efdc6db3d5b4e303ac771bed19d68748.