Sink interface change: put, remove args: (sub, obj) => (obj, sub)
Towards #301. 1493c2e91c2d9763e6023fafbe3d8fedbfba81e5 changed the
`Sink.put()` and `Sink.remove()` to accept a subscription object as the first
paramter. The idea was to make these methods closer to the `Listener`
interface. Subsequent discussions built a consensus that similarity to
`Listener` is less important that putting the most-used paramter first (and
not breaking existing consumers of `Sink`). This change effectively reverts
the "change the Sink interface" parts of
1493c2e91c2d9763e6023fafbe3d8fedbfba81e5.