SyncDAO/VersionNoDAO to use generated "Versioned" classes
This change generates "versioned" classes with properties `version_` and
`deleted_` for interal use with `SyncDAO` and `VersionNoDAO`. Currently,
`VersionNoDAO` expects a versioned `of` (and does not automatically create
one, but maybe it should?).

The value of this change is that client code outside of the `SyncDAO` can
operate on classes that do not contain versioning information; this becomes
an implementation detail hidden inside:
(Sync DAO ==> ... VersionNoDAO => ... ==> <storage mechanism>).