Sync dao fix (#597)
`SyncDAO` was not passing its stress test. Fixing the test code basically required implementing (in the test code) the `AdapterDAO` that @jacksonic requested in my last PR. So, in addition to a minor bug fix in `SyncDAO` this pulls unversioned/versioned adapting *out* of `SyncDAO` and uses new `AdapterDAO` for that.

I also fixed a logic error in `SyncDAO.delegate` management. On init, the delegate should be interpreted as a (potentially stale or unpopulated) cache, not a collection of records yet-to-be-pushed.

* Fix SyncDAO and its stress test

This is achieved by making "versioned class adapting" the responsibility of
new foam.dao.AdapterDAO. Diff of `test/node/syncDAOStress.js` provides
examples of how to correctly compose `AdapterDAO` and `SyncDAO`.

This change yields a `SyncDAO` implementation that passes `syncDAOStress`.

* SyncDAO: Treat delegate as cache (do not push its records on init)