foam.box.node.ForkBox: Support app module loading and secure parsing
This change deprecates `ForkBox.childScriptPath` (but does not yet remove
it). Instead, `ForkBox.appModule` names a script to require() to load app
code. The `appModule` is expected to export a `foam.box.Context` to be used
for IPC with the forked process. The `Context`'s `creationContext` is used on
the parser that reads bootstrap data from the forked process's `stdin`.