Use foam.lookup() for core services in (potential) box contexts
A `foam.box.Context` that uses a `foam.box.ClassWhitelistContext` should not
be forced to whitelist core service classes (that do not need to be
(de)serialized and transported) in order to work. This change eliminates
`foam.net.node.Socket` and `foam.json.Outputter` instantiations that are
often invoked in a `foam.box.Context`. These instantiations now use
`foam.lookup()` instead.