Rework foam.net.node.Server and friends (#783)
New design discussed here: https://groups.google.com/forum/#!topic/foam-framework-discuss/kaFVJVeW9Ek

This is a breaking change for the `Server`, and `*Handler` interfaces.

A `Server` now has one main `Handler` (and a secondary for errors). Common `pathname`, or "route"-based handlers implement `Router` (which extends existing `Handler` interface). Routers also provide inversion-of-control to Handlers being bound to a route. This allows Handlers to decorate themselves, change state, and/or recontextualize when being bound to a specific route in a specific context.