loginByEmail(): Remove email/password.isValid() checks.
These don't increase security.  We don't need to prevent sign in to
accounts with weak passwords or "invalid" emails, we only need to
prevent creating such accounts.

If I manually edit the journal to set a weak password I should be able
to log in with that.  For testing I want an account with admin/admin
as its credentials.