WebModelFileDAO: Add .latch()
Allows injecting models at runtime into the modelDAO so they are found
later.

This could have been .put() but .put() takes fully constructed
objects, and the references in a block of JSON may not be resolved by
the time someone wants to .put() it into the modelDAO, so we use
.latch() to tell it that a model exists that you might want to create
later.