Java Generation: Add foam.core.Identifiable.
If a class has an "id" axiom then we can say its identifiable.  We use
this heuristic in JS all the time but in Java we have to be a bit more
explicit about it.

So here we're adding the Identifiable interface which any object that
has a primary key should implement, and we alias the "id" property of
an object to "getPrimaryKey()" in java.

With this we can clean up a bunch of the primary key code in abstract
DAO and other places.