Implement FirestoreDAO and an authentication adapter
The DAO works with both NodeJS `@google-cloud/firestore` module and the
Firestore web API (v4.9.1). `AuthenticatedDAO` gates API calls to its
delegate on an authentication ID. This does not enforce access, of course,
but does allow DAO operations to queue until authentication is complete.

The `Default*` classes are *not* inner classes so that they are serializable.