Create PublicUserInfoDAO
This commit introduces a new generic DAO decorator that decorates the
find and select methods of a DAO. It takes two property names, one that
holds a user id and one that will hold a subset of the properties of
said user. It looks up the user associated with the id and sets the
other property to an object with a few public properties of the user
that was looked up.

See the documentation in the code for example usage and more details.