Make Element.select() repaint on DAO reset
Currently, the part of the view created by a call to `.select()` will
only update when the dao that it uses is put to or removed from. This
commit changes that behaviour so that the view is updated on any change
to the DAO, not just put and remove specifically.