Fix MDAO test
The MDAO test was getting a beanshell interpreter error because in the
test a call was made to `MDAO.addIndex()`, but the definition of that
method takes varargs, which aren't supported by the version of beanshell
we're using (I think). According to

  https://github.com/beanshell/beanshell/issues/23

though, varargs are now supported in beanshell, so we should update in
order to avoid problems like this in the future.

This commit fixes the problem with a workaround by calling a different
method on MDAO to add the index that doesn't take varargs.