Add Notification Templates.
If a Notification’s template field is populated, then properties of a predefined Notification from the NotificationTemplateDAO will be used to populate the in-flight notification.

SendEmailNotification also now supports sending an email without a template, when the emailName property is empty.
SendSlackNotification will use the Notification body if a slackMessage property is not provided.