Split up the createChildMethod_ refinement.
Java and swift need refinements on createChildMethod_ to merge method info from
parent. There's a subtle difference between swift and java when checking
hasOwnProperty. It's up for debate which platform's behavior is more
correct and once we've decided on that, we should try to get the two
platforms to match.