Fix ProxyListener in JS and Swift
In ProxyDAO always decorate the listener in listen_ rather than listen() as we had before.

The previous behaviour would break if you had several ProxyDAOs in a chain and chaned the delegate of one half way in the chain, listeners against the top proxydao would not get notified.

We need to install a proxy listener for every proxydao in the chain, not just the top one.

Fixes #1381