Fix bug in GroupBySearchView
The bug is that when hardData is set to a falsey value, it isn't saved.
This commit fixes that by checking that hardData is undefined, not just
falsey, which could be a valid value for it to have.

For example: lazy = false would set hardData to false, which is a valid
value for it to be. However, the weak check to see if it was set would
treat it as unset.