Fix bug in IntegerSearchView
There was code left over from CurrencySearchView that multiplies the
number by 100. That needs to be done for currencies because we transfer
and store currency values as integer amounts of cents, but we input them
like floats (ie: with a decimal point).

This commit removes that code and fixes the issue for normal,
non-currency search view inputs.