Modify element lifecycle so that views can be .load()ed after .unload().
This makes views re-usable rather than having to construct an entire new view even time they arrive at .unload().

This add .whileLoaded(f) to Element, which can be used to attach a function that will fire every time a view is loaded.  This is useful for things like DAOListeners that we want to unload when the view is inactive but want to be able to re-attach the next time the view is rendered.