Remove copyFrom warning
We get this warning hundreds of times in the console. Having so many
warnings makes all warnings useless, because no developer is going to
look through hundreds of warnings to see if the code they changed
generated a new warning.

The intent of the warning seems to be to inform the developer that when
they call `copyFrom`, it won't copy anything other than a
`foam.core.Property`. Please correct me if I'm wrong. I think a console
warning isn't the appropriate way to communicate that information to
developers. A documentation comment would be sufficient, or they could
just read the code.

I think the intent of the warning is fine, but the problem is that the
vast majority of the warnings are generated by foam internal code, not
developers *using* foam, who would actually benefit from the warning.