Don't render top nav until logged in
Currently the top nav loads even when the user isn't logged in. This
means that the initE method for child views of the top navigation run
before user is set in the context. This makes it so that the views in
the top nav don't load until the user is properly set in the context.