Fix bug preventing users from adding other users to groups
When you create a group and want to add someone to it, you must have
permission to update that user and you must have the update permission
for the group that you're putting them in. The previous code checked the
group they were currently in, not the group we were putting them in.