Create AuthorizationDAO
This is a DAO that lets models implement their own methods for custom
authorization logic and runs those methods when appropriate.

It's a polymorphic approach to authorization rather than writing many
custom authorization decorators that have to worry about the class of
the objects they're trying to check.

On top of that, it moves authorization code onto the model and removes
the need for a lot of extra files, namely the custom authorization DAO
decoraters we have.