Fix bug when acting as another user
When acting as another user, the user in the context is updated such
that the group is set to the group on the junction between the agent and
the entity. In the checkPermission method in the auth service, we were
looking up the user in the DAO and checking the group of the looked up
user, but that lead to the wrong group being checked. When an agent is
acting as an entity, we want to use the group that was set for the
agent's relationship with the entity, not the agent's usual group.