Fix bug where read authorization check was being called for deletes
It used to call `authorizeOnRead` because calling `this.select_` with
the AuthorizationSink would call `authorizeOnRead` in the `put` method
of the sink. Then the RemoveSink would call `remove_` which would call
`authorizeOnDelete`, which is the call we want.

Now it doesn't call `authorizeOnRead`, just `authorizeOnDelete` like it
should.

It's important that we call `select_` on the delegate instead of `this`
because otherwise we'd still be hitting the AuthorizationSink which
calls the `authorizeOnRead` method.

It's an optimization to pass the delegate to the RemoveSink instead of
`this` because that way we avoid unnecessarily calling
`authorizeOnDelete` a second time when the RemoveSink calls the
`remove_` method on the DAO given to it.