Rework relationship initialization
Separate source vs target initialization.  This provides two benefits:

1) Some classes can be more lazy, as initializing User doesn't necessarily require initializing Group.

2) Complex circular dependencies work more reliably now.  There was a case of

class User;
class SubUser extends User;

class SecurityRecord;

User related one to many to SecurityRecord;
SubUser related one to many to SecurityRecord;

If the SubUser class was asked for, we would hit a circular issue of trying to initialize the relationship before SubUser is finished being built.  This is because building SubUser would ask for User which would as for SecurityRecord which would initialize the SecurityRecord relationships, which would ask for SubUser.