New RoutingJournal Approach
This approach re-uses the existing FileJournal and JDAO with slight modifications.

Instead of writing a single special Journal class that all writes/reads should go thread, we let each DAO have its own FileJournal but structure them so that they write the same underlying file/writer.

We ensure that each journal event is written in full by synchronizing journal operations on a lock, this lock is shared between all journals that write to the same file.

For replay, each DAO will walk the entire journal on its own, skipping any entires that are not tagged for it.  This approach is slower than if we read the whole file in one pass and dispatched events to each related DAO, but it also makes ordering easier to deal with, doesn't require modifing the journal interface and lets the different DAOs be lazily loaded when they are actually needed.  While not ideal, this performance hit is only a one time cost during startup.

A later revision can look at buffering journal events or replaying all the writes in a single pass.