Add agent id to session
Without the agent id, there's no way to tell the difference between two
sessions where user A and user B are both acting as user C because only
the id of user C is saved on the session. This makes it impossible to
query the sessions where a particular user is an agent.

Originally I thought we could just set the userId to the agent's id, but
that would break a large number of places that expect the userId on the
session to be set to what it's currently set to.