Clustering via DAO client/server partially working.
Primary put/remove update both primary and secondary.
Secondary put/remove only update secondary.

Moved order of ClusterDAO in EasyDAO so invoked during replay.

On NSpec creation in factory, walk delegate chain to test and set based on implemented interfaces such as NSpecAware.  Previously if EasyDAO was followed by an OrderedDAO then the NSpecFactory interface assignment would fail because the developer was expecting the NSpect to be the EasyDAO, rather than a wrapper such as an OrderedDAO.