Revert previous changes to email services
In commit 8868279cc998c73e27b745263783c02c57c26191 I made changes to
access the group in the context rather than the user's group. I didn't
realize that these places in particular still should reference the group
on the user being passed in rather than the user in the context. The
change I made in that commit caused the wrong email template to be used
because it was now based on the group of the user doing the put, not the
user who's getting an email generated for them.

This commit reverts those changes to exactly how they were before.