Fix bug in Action
When `maybeCall` is given no data argument, then `isEnabledFor` was
throwing an error because it can't read the 'slot' method of undefined.
This commit fixes that by using `foam.Function.withArgs` instead, which
won't access anything on `data` unless the `isEnabled` method has
parameters, which it doesn't by default.

This came up when adding an Action to a RichChoiceView that doesn't have
an `isEnabled` method and doesn't take any data either. It would fail
when clicked because the call it makes to `isEnabledFor` was throwing an
error.