Fix validation logic
There were errors being thrown when because validateObj was sometimes a
function and sometimes an array. Errors were also being thrown because
sometimes `this.__context__.data$` was undefined. This commit fixes both
of those issues.

It also adds a call to `onDetach` to make sure the listener is destroyed
when appropriate.