Handle/ignore promise rejections in Promised postSet handler
The original code introduced a promise which wasn't returned, and
which if rejected would cause an unhandled promise rejection. The new
code is conceptually similar to this:

```js
function demo() {
  const p = Promise.reject(new Error('demo'));
  p.then(() => {}, () => console.log('handled rejection'));
  return p;
}
```

Calling just `demo()` without doing anything with the return value
will now unfortunately not result in an unhandled promise rejection,
but `demo().then(() => {})` or `await demo()` fortunately would, so if
one actually uses the returned promise errors aren't silenced.

This caused unhandled promise rejections in the confluence test suite.