Get ready for migration to relationship for permissions
This is the last commit before the migration script should be executed.

The migration script will copy all existing permissions off of existing
groups and into groupPermissionJunctionDAO. At this point, all of the
permissions will exist in two places.

After this commit we'll have another commit renaming permissions2 to
permissions and removing the permissions Array property on Group. That
commit will also change all the references to the old permissions
property in the repo and update them to use the new property, which is a
ManyToManyRelationshipImpl, not an Array.