Change calls from getDAO to getJunctionDAO for permissions relationship
This is a performance optimization because we avoid a second lookup in
the permissionDAO only to use the id, which we already had.

It also avoids situations where you don't get all of the permissions you
expect for a group. This happens when you create a junction between a
group and a permission for a permission that doesn't exist in
permissionDAO then call group.getPermissions(x).getDAO().select() to get
all permissions in that group. If any of the permission ids on the
junctions don't have Permission objects in permissionDAO, they won't be
included in the result (and silently at that). By just selecting the
junctions instead and using the permission id on it whether the
permission exists or not, we avoid this situation.

This approach might be controversial. I originally intended to keep it
as is and always require the Permission to also exist in permissionDAO.
Then I had a decorator on the groupPermissionJunctionDAO that would add
permissions to permissionDAO if a GroupPermissionJunction was created
with a permission id that didn't exist in permissionDAO yet. This would
prevent the situation described above and not require developers to have
special knowledge of using the junctionDAO rather than just selecting
from the permissionDAO via the ManyToManyRelationshipDAO.