foam.build.Builder: Enable outputDefaultValue for now
Due to not outputting the default value when building a single foam-bin.js binary, it makes it impossible to disabled a label on an Action

{
  name: 'someAction',
  label: '',
  code: ...
}

When built this will be output as

{
  name: 'someAction',
  code: ...
}

because the empty string is the default value for the label.

This also adds 30-50k to the payload for a large application, so we should find a more proper fix for this.

Perhaps labels should be disabled as label: null, or we should check hasOwnProperty instead of isDefaultValue.

The caveat with hasOwnProperty is that it will output the default value for things like array properties defaulting to the empty array.