Fix major memory leak in Boxes
The response to every single RPC was being held in memory forever.

This change introduces a OneTimeBox which decorates the
replyBox of an RPC request.  We also introduce an ExportBox
which represents a combination of an local box to deliver a received
and a messenger box to given to others which should deliver messages
back to the local box.

Refactored LocalBoxRegistry to use ExportBox instead of the previous
"Registration" class.

Finally updated code to do the registration in the Stub method, rather than
in places like WebSocketService.