Fix privilege escalation session bug
There's a bug in the current code that escalates anyone's privileges to
superadmin if their remote host changes (ie: if they change networks).
The SessionServerBox is using the same session id but replacing its
context with the system context, but with the session in it. This causes
the privilege escalation. The system context has superadmin privileges
and therefore the user's session gains those privileges.

One solution would have been to null out the user, group, and
AuthService cache like we do when creating a new session a few lines
above, but it makes more sense to delete the session outright and force
a login instead. This is explained by the comment in the code.