Provide a single, correct way to update a context based on a session
Adds an 'applyTo' method to Session that serves as a central way of
appropriately updating a context based on a session. A number of places
throughout the repository are updated to call this method to update the
session context instead of doing so manually to ensure consistent
behaviour across the board. This method also allows us to move some of
the logic out of SessionServerBox and into the method, making
SessionServerBox simpler.

This commit also updates the logger for sessions so that it shows both
the agent and the user when a user is acting as another user. This will
make things easier to audit.