Update session object-level access control logic to support SPIDs
This commit makes it so that a user must pass one of three checks in
order to operate on a session. The checks are as follows:

  1. The user is the "session user", meaning the user's id matches the
     value of the userId property on the session instance.
  2. The user has a permission of the form "session.<operation>.<spid>"
     where operation is "create", "read", "update", or "delete". This
     permission allows them to operate on all sessions where the user
     in the session has a SPID that matches the SPID in the permission.
  3. The user has a permission of the form "session.<operation>.<id>"
     where operation is the same as above and id is the id of a session,
     which is a GUID.

The first rule allows users to modify their own session in so far as the
property-level access control checks allow them to. The second rule
allows "weak" admins - which are administrators of service providers - to
manage the sessions of users under their service. The third check allows
"super" admins to be able to do anything. Note that the second check
won't always pass for super admins because it will fail if the userId of
the session is not set.