Add createMode, readMode, and updateMode to Property
These new property hints can be used to specify the view mode of the
property under different controller modes. This opens up the possibility
to hide properties when creating an object but allow it to be edited
when the object is being edited, for example. It allows for more
fine-grained control than simply setting visibility, which should be
considered deprecated once this commit is merged.