Fix authentication bug in AuthWebAgent
The buggy behaviour is that if you access a service that is decorated
with AuthWebAgent and you supply a bearer token, then AuthWebAgent
checks if the _user_ of the session has loginEnabled set to true. This
is incorrect. It should have been checking the agent if it was set and
the user otherwise.

However, checking loginEnabled here is arguably incorrect anyway, since
the sign in process should have caught that. By the time you have a
session with user id (and possibly agent id) already set, you've already
passed the authentication checks. Therefore this commit removes the
check on loginEnabled rather than fixing it.

It also fixes a bug where the wrong group was being checked if the user
was acting as another user. This is correctly implemented in the
Session.applyTo method so we simply call that method instead of doing it
ourselves.

This commit also adds a check to see if the group is null.