Fix bug where PropertyExpr view would crash
The view would throw an error if an invalid StrategyReference were
returned from a call to Strategizer.query. After I had written the view
for PropertyExpr.model, I ended up adding StrategizerChoiceView, which
does the same thing, but also gracefully handles invalid
StrategyReferences by logging a warning and continuing rather than
throwning an Exception.

So changing the view to StrategizerChoiceView has two benefits, it fixes
the crash and it simplifies the code.