Let user select class to create in DAO controller
This commit changes the default detailView for the DAO Controller to
FObjectView instead of DetailView. This allows the user to select the
class that they want to create an instance of.

Since FObjectView's class dropdown is populated by the Strategizer, one
can add StrategyReferences to populate the dropdown. If there are no
alternative strategies for the model being created, then the field to
select the class will be hidden and the base class will be used, (the
'of' of the DAO being controlled), which is consistent with the current
behaviour.