Fix bug where objects couldn't be updated via DAO Controller
High Level
==========

I had only tested create yesterday, not viewing or updating, and
because of that I didn't realize that my changes had broken viewing and
editing via the GUI.

This commit fixes the issue.

Technical
=========

Now we only use FObjectView in the create controller, not the update
controller. This allows users to select the subclass when creating a new
instance, but not when editing an existing object.

To respect the DAO Controller's feature of a custom detail view, this
commit updates FObjectView so that it supports changing the view used
for the FObject. In DAOCreateControllerView we set that view to the
custom detail view for the DAO controller if it's available.