Fix bug when changing class in create controller
When creating a new object, I might change between different subclasses.
The current behaviour of FObjectView is to copyFrom the old object to the
new one when the class changes, but this causes an often unwanted side
effect where default values of the first class chosen are copied to the
new class chosen instead of using the new class' default values.

This commit gets rid of that behaviour and simply creates a new instance
of the new subclass when chosen.