Refactor sections a bit to resolve a few issues
There are a few issues with the current implementation of sections and
detail views that extend AbstractSectionedDetailView that this commit
seeks to remedy.

The first issue is that sections are being rendered even if there are no
visible properties or actions in that section.

The second issue is that, due to a recent change in the way
SectionedDetailView is implemented, sections would be rendered side by
side when they should have been taking up the full width of the parent.
This issue is caused by an extra div being inserted when we return an
element from a slot in detail views. This extra div breaks the CSS grid
and thus the positioning.

The first approach I took to fix the second issue was to use an
ArraySlot of all of the sections and do the forEach inside of the slot's
return value instead of doing a slot of each section individually inside
of the forEach like it currently does (where is where the extra div is
being introduced). This approach had a downside, however. The detail
view would visibly 'flicker' while the slots were re-rendering. This
secondary issue was happening because slots can be hidden if the user
doesn't have permission to see them. Either we hide them by default and
show them if the user does end up having permission, or we show them by
default and hide if the user doesn't end up having permission. Either
way, the view changes. When several sections are doing this at roughly
the same time, the effect is jarring.

The approach taken to fix this secondary issue is to calculate the
visibility of all sections up front while the view is hidden, then show
the view when we're done waiting for all of the async permission checks.
This approach has a few other desirable properties, listed below.

* We can keep track of the properties and actions in each section that
  are visible based on the current 'data' and 'controllerMode' and get
  the detail views to only try to render those properties and actions.
* We also fix the first issue, because while we're calculating which
  sections are visible, we can check if there is at least one visible
  property or action in that section and not include it in the list if
  there isn't one.
* The "interface" that views that extend AbstractSectionedDetailView
  need to work will becomes simpler. In initE, all you need to worry
  about is putting everything inside a slot of `sections`.
  AbstractSectionedDetailView will make sure that `sections` only
  contains visible sections, and will set the properties and actions
  arrays for those sections to be only those that are visible. It does
  this properly by recalculating whenever `data` or `of` changes. This
  removes the need for the views extending AbstractSectionedDetailView
  to use slots of `data`. The slot of `sections` is a slot on `data`
  implicitly.