Fix infinite loop
The slot returns a promise.

ExpressionSlot will detect that its value was set to a promise and will
set its `promise` property to that value. When the promise resolves, it
will set its `value` property to the value that promise resolved to.

In some cases, the `then` callback of the promise in question simply
returns, which means it resolves to undefined in those cases.

Therefore, ExpressionSlot tries to set `value` to undefined in those
cases.

When a property is set to undefined, `clearProperty` is called, which
causes a `propertyChange` event to fire.

There must be something that's listening for that event that ends up
causing the original slot in question to fire again, thus causing an
infinite loop.