Avoid naming conflict between permissions
A group with "user.read.*" will be able to read all users, as intended,
but that permission also implies "user.read.lifecyclestate.deleted" (and
rejected), which is unintentional. We need to be able to give a group
access to read all users without letting them read those that have been
marked as deleted or rejected.