Fix bug where SubSlot was detaching unnecessarily
SubSlots were detaching whenever the class of the object in the parent
slot changed. The reason for this, according to a comment in the code,
is that the new class will have "different SubSlots", which I assume
means different axioms.

An unfortunate side effect of this "feature" is that when the class of
the object in the parent slot changes to another class with the same
axiom, the SubSlot detaches anyway. It would be preferable, in my
opinion, if the SubSlot stays alive as long as the new class has the
same axiom the subslot was set up for. That's what this commit
implements.