Fix phone number validation when not required
Before this commit, if a PhoneNumber property was not required: true,
then validation wouldn't be applied at all, even if the user typed
something into the input. Instead, we want an empty input to be valid,
but we still want to use the normal validation logic if a value is
entered.