Make ChoiceView use default value when user selects placeholder
Otherwise the user doesn't have any way of explicitly choosing the
default value.

The important part is that it's an _explicit_ action taken by the user.
When we were setting `this.data` to `undefined` before, we were telling
the data to go back to an "untouched" state. This has the undesirable
side effect of copyFrom ignoring that property since it's in an
untouched state. Instead, we want to explicitly set that property to a
default value (such as 0 for a long property, or "" for a string
property) so that copyFrom will copy that default value over.