Don't filter out PENDING objects by default in LifecycleAwareDAO
Doing so makes it impossible to view the object being approved for an
approval request. Instead, we allow PENDING objects to be included in
results when `select`ing or `find`ing and leave it up to the call site
to include the a predicate to filter out PENDING items if desired.