Make table refresh button less prominent
It's not destructive, so it shouldn't be styled as destructive, and it's
not the primary action of the DAO controller, so it shouldn't occupy the
same space as the create button. This commit moves it alongside the
export button as a secondary action. I think it's more appropriate there
since it's available for any table and is always a secondary action,
just like exporting.